Remove NodePoolIndex from Node
Currently nodes have a global index (a globally unique identifier
number) and a pool index. The global index is what optimizations and the
compiler are supposed to use to track the identity of a node since it is
unique for the life of the compile. A notion of being able to garbage
collect nodes was attempted in the node pool and this involved the use
of a pool index. This index is now dead and should not be used. This
commit removes the field and cleans up the associated infrastructure.

Signed-off-by: Andrew Craik <ajcraik@ca.ibm.com>