Update Tril-based opcode tests
Previously, opcodes were specified as arguments to Tril-based opcode
tests by their name but without the type specifier (e.g. `add` instead
of `iadd`). This is a confusing and make the code more difficult to
understand. This commit therefore changes the tests to take the full
opcode name as argument.

Also, according Google Test guidelines, death tests should have names
where first part of the name ends in `DeathTest`. This is to ensure
these are run first in potentially multithreaded environments. This
commit updates the compiler death tests to conform with the guideline.

Signed-off-by: Leonardo Banderali <leob@linux.vnet.ibm.com>