Convert LoopStrider _storeTreesList to map
The _storeTreesList data structure is currently an eagerly allocated
data structure mapping symrefs to a ScratchList of store tree tops. This
eager allocation is overly aggressive and it makes much more sense to
create a map from symref numbers to lists where entries are initialized
as needed. This saves memory and compile time.

Signed-off-by: Andrew Craik <ajcraik@ca.ibm.com>