Fix segfault when running omrgctest with options
When initGoogle is called it will remove the arguments it recognizes from
argv and decrements argc accordingly.

The seg fault was occurring because argv was being updated while argc was not. This is because argc is used to initialize _argc before initGoogle is called. Generally initGoogle should be the first thing that is called.

Instead of fetching the testcases from the config file passed into the command line, the test now already knows which testcases it will be running. To choose which set of sets to run the --gtest_filter="" option can be used.

Issue:#1278

Signed-off-by: LinaSerry <linaserry@cmail.carleton.ca>