Add TypedAllocator to jitbuilder/release/include
TypedAllocator will be used with C++ Standard Library containers
and classes such as std::basic_string and std::map, in order
to use our custom allocation mechanism. It needs to be included
in MethodBuilder and TypeDictionary headers in order to allow us to
declare std::maps and std::basic_strings which use it.

Signed-off-by: Sajid Ahmed <Syed.Sajid.Ahmed23@ibm.com>