Delete TypedDictionary copy constructor
Implicit generation of the copy constructor will lead to ambiguous
compiler errors (in external compilation units) when using
TR::typed_allocator for the STL containers and std::basic_strings
in TypeDictionary. Since it does not make sense, both from semantic
and functional perspectives to copy a TypeDictionary, its copy
constructor is explicitly deleted. Necessary changes have been made in
the JitBuilder examples to avoid the usage of the copy constructor.

Signed-off-by: Sajid Ahmed <Syed.Sajid.Ahmed23@ibm.com>