Use OSR method info to check for OSR blocks
IL (intermediate language) generation sometimes assumed that the IL for
a method has OSR (on-stack replacement) code and catch blocks whenever
getOSRPoints() is nonempty for that method. However, it's possible that
all OSR points for the method belong to other inlined call sites, since
the set of OSR points can differ when IL is generated for a single
method repeatedly. For instance, symbol references that are unresolved
when first inlining a method may be resolved when later inlining the
same method again.

This commit changes IL generation to determine the presence or absence
of the OSR code and catch block based on the OSR method info for the
current inlined site index, without consulting getOSRPoints().

Signed-off-by: Devin Papineau <devinmp@ca.ibm.com>