Convert RegisterCandidate bool fields to flags
The RegisterCanddiate class currently has a number of bool fields which
hold flags. These are allocated 1 byte each by the compiler when they
can fit in one bit. This changes creates a flags field and associated
enum and moves the storage of all flags into the field to save space.

Signed-off-by: Andrew Craik <ajcraik@ca.ibm.com>