Compact RegisterCandidate block information
Currently RegisterCandidate block information stores a block number and
a load/store count for every block where a candidate is being
considered. There are a number of zero entries added to this data
structure and these zero entries are expensive in terms of memory
footprint. This change adds a TR_BitVector to track the candidate blocks
allowing the map to contract to only store blocks with a non-zero
load/store count to save space.

Signed-off-by: Andrew Craik <ajcraik@ca.ibm.com>