Correct JitBuilder opt strategy opt order
Current ordering has induction variable analysis going before
loop canonicalization. Sadly, loop canonicalizer changes the
structure of the loop but does not manage to update all the
induction variable info, so wrong blocks are pointed at which
can lead to loop unroller crashing (we don't like that).

Swap the order so that loop canonicalizer goes first and
then induction variable analysis can proceed on the already
transformed graph. While debugging this problem, I noticed
that the blocks going into this process can be in a wacky
order which a pass of block ordering can clean up nicely.
So I also added that in.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>