Fix assertion in OperandArrayTests
This commit fixes a level of indirection problem with OperandArrayTests to
avoid a recently added assert in IndexAt that requires its first
argument to be of type TR::Address. This problem was not noticed because
OperandArrayTests is not currently run as part of our PR testing
because the test doesn't work on all platforms. That will be addressed
by a separate commit.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>