Test jitbuilder testall on linux_x86,osx for PRs
Right now, we only run the "make test" set of JitBuilder tests on
linux_x86 and osx as part of our pull request testing, but we can
actually run "make testall" which runs more extensive testing because
these platforms support those tests.

We recently missed a problem in one of the "testall" tests because we
aren't running them, so it's not just a theoretical benefit.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>