Fix child handling in SIMDsplatsEvaluator on x86
The child of a vsplats node can return either a TR_GPR or TR_FPR register
on evaluation. Only the TR_FPR register case was being handled correctly.

This change adds a MOVD / MOVQ instuction to transfer the data from the
TR_GPR register into a TR_VRF register which is compatible with the
pshufd instruction used to perform the actual splating.

Issue: #1465
Signed-off-by: jimmyk <jimmyk@ca.ibm.com>