Convert storeTreesAsArray to a deque in CopyProp
Copy propagation currently builds up a list of store trees which is
stored in a single block of stack allocated memory. Doing this requires
walking all the trees to first find the number of nodes that need to be
stored and then storing the treetops after allocating memory. Using a
deque saves one full walk of the trees.

Signed-off-by: Andrew Craik <ajcraik@ca.ibm.com>