Free valgrind objects through addFreeMemory
Free memory in MM_SweepPoolManagerAddressOrderedListBase::addFreeMemory. This avoids using valgrind APIs in exampleVM and uses a new valgrind API by Andrew.
There are still a few errors because of slot objects that need to be fixed.

Issue: #826
Signed-off-by: Varun Garg <varun.10@live.com>

Improved valgrind calls in addFreeMemory

Remaining valgrind issue is with the calls from final else condition - "/* Check if the hole is a free list candidate */ ".
If we skip valgrind call in that block, we get errors of a few overwritten objects.
The overlapping errors are still less than before, meaning the other valgrind calls are ok in the class.
If we add the call, we get errors due to inaccessable slot objects later.

Issue: #826
Signed-off-by: Varun Garg <varun.10@live.com>