Transform GlobalDSE to save compile time
The current global dead store elimination optimization implementaiton
has a structure which processes all use and def indices for each
particular node using data structures sized based on the number of
usedef inidices. This optimization changes the implementation to use
only the defs and uses found in a region to a more compact check saving
both compile time and space.

Signed-off-by: Andrew Craik <ajcraik@ca.ibm.com>