Short circuit single block loop check in GlobalDSE
Global dead store elimination currently has a check for single block
loops which involves gathering all the blocks in the region structure.
This change adds a short circuit check that the region structure being
checked has, itself, only one subnode (either region or block) since
having more than one automatically excludes the region and the check is
essentially free compared to the cost of gathering the nodes in the
region.

Signed-off-by: Andrew Craik <ajcraik@ca.ibm.com>