Improve JitBuilder handling of shift operands
After the new IL validator showed up, it exposed some weaknesses in
how operands of shift operations were being handled by JitBuilder.
In particular, shift operands are validated using the same mechanism
as for other binary operations, and that mechanism is entirely
inappropriate for shifts. In fact, it only really works for 32-bit
shift operands

This commit separates out the handling of shift operations from the
generic binaryOp* handling and makes the operand testing a bit more
relevant: the left operand must be integer typed and the right operand
will be automatically converted to a 32-bit integer. If no conversion
to Int32 exists for the right operand's type, ConvertTo should assert.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>