Fix _patchVirtualGuard to not assert on false positives
When patching happens synchronously we are effectively patching in a
jump with a guarantee no processor will be executing the patch location.
Thus there is no race condition. This means we will be overwriting
existing instructions. If it so happens that the instruction at the
patch location has opcode 0xA7 or 0xC0 as the first byte we may
encounter a false positive and incorrectly assert. Such an example can
be seen if an LHI R8,0 (0xA7880000) instruction is after the VGNOP.

This fix strengthens the condition of when one byte patching happens,
i.e. if and only if we are certain a BRC or BRCL is present.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>