Ensure assembly conversion routines see all src
The compiler sources are split into two portions: The source provided by
the core compiler technology, and the source provided by the consumer of
the compiler technology.

Previous to this patch, only files provided by a consumer of the
technology would be consumed by the pasm2asm and masm2gas functions.

Signed-off-by: Matthew Gaudet <magaudet@ca.ibm.com>