Move ASTValueType enum into ASTValue and rename it
This reflects the fact that the type information belongs to the class.
It now also makes sense to rename the enum from ASTValueType to ASTType.
Lastly, it also makes sense to give the enum members more appropriate
names. The reasoning for this is that the AST does not care about the
underlying representation of values. It only cares about whether the
value is an integral, a floating point, or a c-string.

Signed-off-by: Leonardo Banderali <leob@linux.vnet.ibm.com>