Disable PartialRedundancyElimination when profiling
When profiling with HCR enabled, the extra blocks
and temps required in this configuration can result
in PRE taking up a significant amount of compile
time and memory. This can often result in the
compilation running out of scratch space during
or soon after PRE, degrading performance.

Disabling this opt in this conditions has been
verified on several benchmarks, with no evident
regression, as the following recompile will still
include the opt.

Signed-off-by: Nicholas Coughlin <cnic@ca.ibm.com>