Reduce OSR shared slot map memory usage
The OSR shared slot map can be compressed by representing
a continuous set of entries that share the same mapping
with the first. This was implemented by walking the
map and removing any duplicates that were identified,
however, TR_Array implements remove by copying all remaining
elements forward by one index.

The new implementation builds up a compressed list as it goes
along, rather than removing in place. This will not always be
cheaper, but will reduce overall usage due to the high frequency
of entries that can be removed.

Signed-off-by: Nicholas Coughlin <cnic@ca.ibm.com>