Fix child handling in SIMDsplatsEvaluator on x86
SIMDsplatsEvaluator on x86 attempted to use a movq instruction when
handling Int64 data on 32bit which is not allowed. Furthermore,
it did not recognize the fact that the first child would come in as
a Pair Register under 32 bit when handling Int64 type data.

This change adds a is32bit check to SIMDsplatsEvaluator when handling
Int64 type data. If true, the evaluator will know that the incoming data
is in a pair register and handle it accordingly. These changes required
the use of a por instruction which was also implemented.

Issue: #1465
Signed-off-by: jimmyk <jimmyk@ca.ibm.com>