Fix lowering methodExitHook simplifier
With inlining in debugging mode, there might be multiple method
enter/exit hooks in one extended basic block and the old way of
lowering them is wrong. This PR collect enter/exit treetops and nodes
into a list and process them after finishing all other nodes in
the same extended basic blocks to avoid spliting a block while in
the process of analyzing the block.

Signed-off-by: Yi Zhang <yizhang@ca.ibm.com>