Fix long displacement for RSL instructions
RSL instructions (such as TP) do not have an index register. As such we
must treat them specially when generating long displacement logic.
Similar logic already exists for various other kinds of instructions
with memory references that do not have an index register.

To prevent future problems in this area we also take the opportunity to
add a fatal assert if this issue arises in the future.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>