Remove assertions in _patchVirtualGuard
It does not appear for there to be a safe way to tell whether we are
patching synchronously or asynchronously without false positives. The
asserts placed in this function are firing off on upstream projects
because runtime virtual guards that are placed as NOP guards could have
arbitrary code after them, even other NOP branches. 

As such for a synchronous virtual guard we will be clobbering the block
it is guarding and we cannot be certain what is at the location we will
be patching. It could in fact be another NOP branch which may confuse us
when trying to distinguish whether we are patching a NOP BRC / BRCL.

All assertions have been removed from this function. Furthermore the
patching order has been re-ordered to be as safe as possible with the
information we have at hand. It will be up to the caller / codegen to
ensure the non-atomic patching is safe.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>