mem-info verbose GC during active CS
At the end of first and at the start of second/last STW phase of
Concurrent SCavenger Cycle, mem-info Nursery stanza will have two parts:
evacuate and hybrid allocate-survivor.

A the end of first STW phase, nursery may look like shown below.
Allocate/survivor is expectedly a smaller part (19%) of total Nursery,
and it is mostly (88%) free, since only objects found through root
scanning has been copied. By the end of concurrent phase,
allocate/survivor as reported at the beginning of last STW phase would
typically have little or no free memory.
Evacuate is always unavailable for allocation and reports 0 free memory.

     <mem type="nursery" free="3247832" total="16777216" percent="19">
       <mem type="allocate/survivor" free="3247832" total="3670016"
percent="88" />
       <mem type="evacuate" free="0" total="13107200" percent="0" />
     </mem>
     
A couple of xxxxxFreeSurvivorMemorySize() APIs in SemiSpace has changed
its behavior. It was previously  reporting hard coded zero (since any
free memory in Survivor is reserved when application runs). Now, the API
will return the true free memory in its memory pool, at the moment of
invocation, and than it's callers responsibility to treat that value as
valid or not (depending if we are in a middle of a concurrent phase or
not).

Also, clear allocation stats at the beginning of Concurrent cycle, so
that allocate-stats stanza at the end of the cycle reports allocation
only during the cycle. This makes _bytesAllocatedAtConcurrentStart in
SemiSpace code obsolete (effectively, it's always 0).

Signed-off-by: Aleksandar Micic <amicic@ca.ibm.com>