Implement reduction of ifxcmpge to ifxcmpeq
In specific cases, we can reduce the ifxcmpge opcode to ifxcmpeq.
Once this transformation is performed, the test-under-mask optimization
can be applied to the ifxcmpeq node.

The reduction is possible on the following tree:

ificmpge --> block_5
  iand
    iload <...>
    iconst 16
  iconst 16
 
Since the constants are equal, the greater-than case is impossible and
ifxcmpeq can be used instead.

Issue: #1556
Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>