Prevent LoopVersioner from checking guards
LoopVersioner will search for profiled nodes,
to build versions of loops where the profiling
is disabled. If a node beneath a nopable guard
shares BCI with a profiled value, LoopVersioner may
mistake the guard's child as being profiled. This
change prevents the method from analysing the children
for a nopable guard.

Signed-off-by: Nicholas Coughlin <cnic@ca.ibm.com>