Support tools cross compilation
This patch lets us compile and use the tools when cross compiling for
x86 on an x86-64 machine.

I've simplified the tools directory, moving the add_tracegen and
add_hookgen macros to seperate modules. This lets them be used outside
of OMR more easily. Now, we can completely skip the tools directory when
importing.

Now, do not skip the tools directory when cross compiling. Instead, emit a
message that the tools will be cross compiled.

The OMR_TOOLS flag is now respected. If youi have the tools disabled,
and are not importing them, the build will error. However, this patch
lays the groundwork for using already-installed build tools.

Signed-off-by: Robert Young <rwy0717@gmail.com>