Stop using target_compile_options for compiler
The issue with target_compile_flags is that there's no portable way to
split do compilation flags differently between C and C++. The
$<COMPILE_LANGUAGE> generator expression was added in 3.3, which is
newer than is avalable on TravisCI, and is explicitly not portable to
Visual Studio.

We need the ability to specify some flags to C files, and others to C++
flags.

Signed-off-by: Matthew Gaudet <magaudet@ca.ibm.com>