Fix resolution of allocate for non-size_t integers
Currently, resolution of TR::Region::allocate for integers that are not
of type size_t resolves to the function template used to instantiate
managed instances of objects. This is dangerous.

To fix this, change the name of the function template to 'create' so
that it is no longer a homonym of the standard allocate function.

Signed-off-by: Luc des Trois Maisons <lmaisons@ca.ibm.com>