Initial ELF Object File Support
Provide the ability to create object files from JIT Builder
compilations. Add the compiler option 'enableObjectFileGeneration' to
guard this functionality.

This functionality is initially only supported on x86-64, and supports
exporting a compiled symbol for inclusion during linking, and can
produce the relocation entries required to support function dispatch via
a load of an immediate into a register + a call to that register.

Signed-off-by: Luc des Trois Maisons <lmaisons@ca.ibm.com>