Merge stats across CS phases
In Concurrent Scavenger gather per phase/increment Scavenger stats.
These are typically used for reported purposes. Than at the end of a
cycle merge those stats. Those are typically used for various heuristics
(like tilting) that are performed at the end of a cycle. Currently,
since we clear per increment stats and do not merge them, the heuristics
rely only on last increment, which is vary small data subset.

Standard Scavenger is not meant to be affected, although a lot of code
is shared with CS. There is still a notion of an increment, but there is
only one. We still merge that one with the cycle stats, so they are
identical.

Signed-off-by: Aleksandar Micic <amicic@ca.ibm.com>