Move addPrefetch code from OMR to OpenJ9
This removes addPrefetch from OMR. It is replaced by a call to the method
insertPrefetchIfNecessary. By default insertPrefetchIfNecessary just
returns. Extending projects can override it to add functionality.

iloadEvaluator and l2aEvaluator were also modified to now use
insertPrefetchIfNecessary instead of addPrefetch.

Prefetch was added to iloadEvaluator on Power because it was shown to
improve performance on an internal appserver benchmark.

The OpenJ9 changes will be handled by a pull request in OpenJ9.

Issue: #2240
Issue: #2634
Signed-off-by: jimmyk <jimmyk@ca.ibm.com>