Support folding immutable array shadows
If we know an array is not going to be changed nor is its content, we
can fold the load of its element. This is done by adding a bit in
KnownObjectTable to indicate whether a known object is an array with
immutable content. When VP sees an array shadow load from such array,
it will improve the array shadow to an ImmutableArrayShadow, and ask
transformIndirectLoadChain to fold it. ImmutableArrayShadow is also
served as an indicator to transformIndirectLoadChain that it is safe
to fold an array shadow with such symbol.

Signed-off-by: Liqun Liu <liqunl@ca.ibm.com>