Hoist mockStrategy support into OMR Optimizer
This functionality allows code external to the compiler (such as a test
case) to specify a mockOptimization strategy that overrides the
optimization strategy a particular method might be expected to use.

This is currently not a thread-safe implementation because doing a
thread-safe implementation would require a more invasive compilation
control modification (since the strategy would need to be injected into
the Compilation object, but due to where it's created, there's little
opportunity for this.)

The good news is that Google Test doesn't test in threads, so for now
we are spared the necessity for dealing with this.

Signed-off-by: Matthew Gaudet <magaudet@ca.ibm.com>