Fix compilation issues on zOS (#1639)
m4 on zOS can not handle C/C++ style comments. This file is included in
a C++ header and a few m4 files so the copyright header needs to work
for both. Remove the copyright header while a solution is being found.

zOS assembler can not handle lines with more than 71 characters even in
comments. When a solution is found ensure that the length of each line
in the copyright header is a maximum of 71 characters.

[ci skip] since this is just removing a copyright header

Signed-off-by: Charlie Gracie <charlie.gracie@gmail.com>