Use vector rather than hash table in CompactLocals
CompactLocals frequently has to find the entry
in its interference graph for a live local index.
This was done with the interference graph's built
in hash table, however, this is expensive.
CompactLocals also had an array of symbols, indexed
by live local indices.

This change replaces the array with a vector of
interference graph nodes, allowing CompactLocals
to cheaply look up the interference graph node
for a live local index.

Signed-off-by: Nicholas Coughlin <cnic@ca.ibm.com>