Correct register lookup when computing available
When determining if a register is a suitable
candidate for a local, GRA will check if the local's
use overlaps with any prior allocations. One check
confirms that when a potential new register allocation
exists at the entry of the block and there is an existing
allocation at the exit, the uses in the block do not
overlap. However, when requesting the existing register
allocation for a block, it would ask for the allocation
against the block entry, rather than the exit.

Signed-off-by: Nicholas Coughlin <cnic@ca.ibm.com>