Rework GlobalSet to allow for BitVector operations
Recent memory improvements modified RegisterCandidate's
BlockInfo, such that blocks with a weight of 0 did
not have to be entered in the map, rather they were
set in a BitVector. Direct access to this BitVector
allows a majority of blocks to be initialized with
BitVector operations, rather than having to consider
every block for every candidate.

To achieve this, GlobalSet had to be flipped from
BitVectors of symbol references accessed in a block
to BitVectors of blocks that access a symbol reference.

Signed-off-by: Nicholas Coughlin <cnic@ca.ibm.com>