Convert Tril tests to be backend agnostic
Addresses #1668, which would like the ability to use Tril to test
multiple compiler backends. This commit adds the notion of
Tril::DefaultCompiler, and opaque concrete Tril::MethodCompiler.

It is selected by a CMake flag, -DTRIL_TEST_COMPILER=1, which uses the
test compiler as a backend, or without, where JitBuilder is used as a
backend.

Signed-off-by: Matthew Gaudet <magaudet@ca.ibm.com>