Refactor double/float remainder helper
Function floatRemHelper is used to generate code for node drem/frem.
Generated assembly uses divide-to-integer instructions DIxBR to get
remainder of two floating point numbers. In rare case there is a
call-out to helper function to get the result. This commit uses
system linkage to call the fast-path helper function in OOL section.

Signed-off-by: Rahil Shah <rahil@ca.ibm.com>