Avoid inflating frequencies in block splitter
When block splitter copies a merge block, it distributes some of the
frequency of the original block to the copy, based on the proportion of
the predecessors' total frequency belonging to the hot predecessor. A
subsequent check raises the resulting frequency if necessary to ensure
that it's no less than that of the predecessor, with a static debug
counter suggesting that otherwise the frequencies would be inconsistent.
However, when the predecessor branches, it's expected for the new block
to be less frequent.

In particular, if the predecessor is in a loop not containing the merge
block, the duplicate's frequency would be raised to match frequencies
within the loop, interfering with proper layout of the loop.

This commit restricts the frequency adjustment to cases in which the
predecessor does not branch, i.e. cases in which it really would restore
consistency.

Signed-off-by: Devin Papineau <devinmp@ca.ibm.com>