Skip blocks without OSR points is OSRLiveRange
OSRLiveRangeAnalysis will maintain liveness
throghout a block, so that it knows the state
at an OSR point. However, blocks that contain
no OSR points have their trees walked for no
purpose. This change detects whether there is
an OSR point based on the presence of an
exception edge to an OSR block.

Signed-off-by: Nicholas Coughlin <cnic@ca.ibm.com>