Fix maxprec() call to avoid fixed-point overflows on Z
The maxprec() routine that obtains current time on Z-plaforms
triggers fixed-point overflows if the PSW Fixed Point Oveflow mask
happens to be turned on. The reason for the overflow is that a subtract
operation used for calculating the current time mistakenly uses signed
rather than unsigned subtraction. The time value of the TOD-clock
(returned via STCK/STCKF instruction) cannot be negative, so a
subtract logical should be used instead, which would avoid
Fixed-point Overflows.

The fix applies to z/OS, z/Linux and z/TPF platforms.

Signed-off-by: Deepthi Sebastian <deesebas@in.ibm.com>