Move ELF test make dependencies
Move the make dependencies for the ELF tests from ALL_TESTS to explicit
dependencies of experimental_goal.

This is required because entries in ALL_TESTS get built on all
platforms, even if they can't run on all platforms. However, the
distinction between 'building' and 'running' the ELF tests if fuzzy
enough that if a platform doesn't support running the test, it doesn't
support building it either.

Fixes #1748

Signed-off-by: Luc des Trois Maisons <lmaisons@ca.ibm.com>