Added VALGRIND_CHECK_MEM_IS_DEFINED check
Request VALGRIND_CHECK_MEM_IS_DEFINED is called before an object is freed to check api integration itself.
Before calling it, object type is checked, weather it is a omr object or a forward pointer and size is set according to it.
Some improvements and fixes in API in Scavanger have also been made to get it right.

Signed-off-by: Varun Garg <varun.10@live.com>