Fix comptest failures on Linux 390
Shifting a 32 bit integer by more than 31 bits is undefined behaviour in
the C++ spec. Currently the OMR compiler  masks the shift value by 0x1F.
This means a shift of 32 or larger on a 32 bit integer effectively
returns the same value. To ensure the C++ version matches the compiled
version of the test the C++ version needs to also mask the shift value.

Signed-off-by: Charlie Gracie <charlie.gracie@gmail.com>