Improve Travis Linux build times
Currently all of the addon packages were being installed for all Linux
builds. Reduce the initial set to the minimum for all builds and then if
a particular build needs more packages specify the packages for that
individual build. This removes ~10 mins of machine time from each PR and
delivered commit.

Signed-off-by: Charlie Gracie <charlie.gracie@gmail.com>