Upgrade brew before using it
Brew is not very smart. When we install ccache, brew will auto update
first, installing ruby 2.3. Unfortunately, the install will require the
new ruby, but will run with the old ruby brew used to upgrade. So now,
we need to update brew before running any other commands. This is a new
issue with brew.

This patch will fix all the brew-related osx travis failures we are
seeing across the board.

Signed-off-by: Robert Young <rwy0717@gmail.com>