Fix types of generated JIT functions
The JIT functions are declared as accepting signed integers and as such
we must declare for unsigned tests that the parameter type of the
generated JIT function is also signed. Furthermore we must ensure the
width of the parameter type is the same as the width of the input as
implicit promotion to at least a 32-bit width will be performed. The
parameter type having the same width as the input ensures the we will
not perform a logical extension and then a type conversion, but rather
a signed extension to a signed integer type as defined by the JIT
function definition.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>