Bridge work to move instruction pointers to CodeGenerator
As part of the work to relocate the management of the `TR::Instruction`
pointers from the Compilation object to the CodeGenerator, introduce
the same functions in the code generator that simply call their
counterparts in the Compilation object.

This is only to help stage the transition effort in OMR and all
known downstream projects.  The appropriate instruction fields
will be moved into the CodeGenerator once all callsites have been
updated.

Signed-off-by: Daryl Maier <maier@ca.ibm.com>