Deprecate OMR::Z::RealRegister::getRegMask
Deprecate OMR::Z::RealRegister::getRegMask in favor of the already
existing API OMR::RealRegister::getRealRegisterMask which is currently
not used in the IBM Z codegen. The base OMR class already has a field
which stores this mask value so we may as well make use of it.

Furhter to this, it has been identified that the deprecated function is
quite hot in a compile time profile of the JIT, likely because of how
often it is called. Caching the mask value will improve performance,
rather than computing it every time via the sequence of if / else
blocks seen in the now deprecated function.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>