Clean up genericLongAndAsRotateHelper
-Rename the function to tryToReplaceLongAndWithRotateInstruction
as this is more descriptive

-Clean up the function to avoid skipping the evaluation of i2l
and iu2l nodes. This should be handled by the simplifier (and
it currently is). Also, insert a static debug counter to catch
potential opportunities in the future.

-Add asserts during binary encoding and instruction construction
to make sure the reserved bits in the instruction's operands are
not being set. This is a source of printing errors in trace logs.

-Test the function using tril + google tests

Signed-off-by: Dhruv Chopra <Dhruv.C.Chopra@ibm.com>