Check subregion exits in checkStructure
Now checkStructure will check the consistency between exit edges from a
region, and edges outgoing from the subnode corresponding to that region
in its parent. If r and p are regions, and n is a node in p with
structure r, then in order for structure to be consistent, the following
condition must hold for both normal and exception edges: For each node
number i, r has an exit edge to i if and only if n has an outgoing edge
in p that targets i.

Signed-off-by: Devin Papineau <devinmp@ca.ibm.com>