Add structure exit extraction
After deleting edges from the control flow graph, there may be blocks,
or structure subgraph nodes more generally, from which execution is
guaranteed to exit a containing loop. A fresh structural analysis would
consider such nodes to be outside the loop, which is preferable for the
sake of later loop optimizations.

Setting aside the general benefit of smaller loops, leaving loop regions
to hold onto nodes that don't belong can cause significant problems,
such as versioning of checks that are guaranteed to fail.

This commit adds TR_RegionStructure::extractUnconditionalExits, which
searches for nodes that no longer belong within loops, and pulls them
out of those loops. This is not done as part of automatic structure
repair, because it can be surprising and unwanted when the control flow
graph is in an intermediate state. Currently this is done only at the
end of value propagation.

Signed-off-by: Devin Papineau <devinmp@ca.ibm.com>