Support specifying the type of indirect vector loads and stores
The typeless vector support in TR needs types specified at certain
points. One such point is the symref of indirect loads and stores. Since
Tril currently constructs its own symrefs, this patch adds a new
property to allow specifying a type vstorei and vloadi. This looks like

   (vstorei type=VectorDouble offset=0 (aload ... ) )

In order to simplify the indirect opcodes a bit, this also eases the
requirements for indirection, making offset optional, but defaulting to
zero.

A test case is included that validates VectorDouble vadd

Signed-off-by: Matthew Gaudet <magaudet@ca.ibm.com>