Migrate nodeNeeds2Regs from Compilation to Node
A node should be responsible for answering whether it requires a register
pair or not.

Move the functionality to the Node class, rename the function to
`requiresRegisterPair`, and add some documentation.

Redirect the Compilation query to the Node query to allow downstream
projects to complete refactoring.  The Compilation shim will be removed
when this has been completed in all known downstream projects.

Signed-off-by: Daryl Maier <maier@ca.ibm.com>